src/dspygen/experiments/cliapi/cli_app.py

__app_name__ = 'GitHub CLI'
__version__ = '2.35.0'
__description__ = '''A command-line tool for interacting with GitHub.'''
__author__ = 'GitHub, Inc.'

import typer

app = typer.Typer(help=__description__)




# Configurations
configurations = {
    'globals': {
        'editor': {'default': 'vim', 'description': 'Default text editor for GitHub CLI operations.', 'type': 'string'},
        'pager': {'default': 'less', 'description': 'Default pager for displaying command outputs.', 'type': 'string'},
        'theme': {'default': 'auto', 'description': 'Color theme for GitHub CLI output.', 'type': 'string'},
    },
    'repository': {
        'default_branch': {'default': 'main', 'description': 'Default branch name for new repositories.', 'type': 'string'},
        'visibility': {'default': 'private', 'description': 'Default visibility for new repositories.', 'type': 'string'},
    }
}





# Voice Settings
voice_settings = {'version': '1.0.0', 'voice_format': 'wav', 'voice_language': 'en'}





# Accessibility Features
accessibility_features = {'version': '1.0.0', 'screen_reader_support': True, 'high_contrast_mode': False, 'keyboard_navigation': True}





gh_app = typer.Typer(help='GitHub CLI main command for interacting with GitHub repositories and services.')

@gh_app.callback()
def gh_callback(version: bool = typer.Option(False, help='Display the version of GitHub CLI.'), help: bool = typer.Option(False, help='Show help information for GitHub CLI.')):
    pass

repo_app = typer.Typer(help='Manage GitHub repositories.')

@repo_app.command(name='clone', help='Clone a repository to your local machine.')
def clone(repository: str = typer.Argument(..., help='The repository to clone.'), depth: int = typer.Option(None, help='Create a shallow clone with a history truncated to the specified number of commits.')):
    '''Clone a repository to your local machine.'''
    typer.echo('Executing clone subcommand')
    # Examples:
    # gh repo clone owner/repo
    # gh repo clone owner/repo --depth 1

gh_app.add_typer(repo_app, name='repo')

app.add_typer(gh_app, name='gh')





octo_org/octo_plugin_app = typer.Typer(help='A plugin to enhance GitHub CLI with additional organizational tools.')

octo_app = typer.Typer(help='Organizational tools provided by the octo-plugin.')

@octo_app.command(name='sync', help='Synchronize organizational repositories.')
def sync(organization: str = typer.Argument(..., help='Name of the organization to synchronize.'), force: bool = typer.Option(False, help='Force synchronization even if conflicts exist.')):
    '''Synchronize organizational repositories.'''
    typer.echo('Executing sync subcommand')
    # Examples:
    # gh octo sync octo-org --force

octo_org/octo_plugin_app.add_typer(octo_app, name='octo')

app.add_typer(octo_org/octo_plugin_app, name='octo-org/octo-plugin')





gh_extras_app = typer.Typer(help='Additional commands and functionalities for GitHub CLI.')

extras_app = typer.Typer(help='Extra tools and utilities.')

@extras_app.command(name='deploy', help='Deploy the repository to a specified environments.')
def deploy(repository: str = typer.Argument(..., help='Repository to deploy.'), env: str = typer.Option(..., help='Target environments for deployment.')):
    '''Deploy the repository to a specified environments.'''
    typer.echo('Executing deploy subcommand')
    # Examples:
    # gh extras deploy owner/repo --env production

@extras_app.command(name='backup', help='Backup the repository data.')
def backup(repository: str = typer.Argument(..., help='Repository to backup.'), destination: str = typer.Option(..., help='Backup destination path.')):
    '''Backup the repository data.'''
    typer.echo('Executing backup subcommand')
    # Examples:
    # gh extras backup owner/repo --destination /backups/repo-backup.tar.gz

gh_extras_app.add_typer(extras_app, name='extras')

app.add_typer(gh_extras_app, name='gh-extras')





GitHub Marketplace_app = typer.Typer(help='A marketplace for GitHub CLI plugins and extensions.')

@GitHub Marketplace_app.command(name='browse', help='Browse available plugins and extensions in the GitHub Marketplace.')
def browse(query: str = typer.Argument(None, help='Search query term.'), category: str = typer.Option(None, help='Filter by category.'), sort: str = typer.Option(None, help='Sort results by criteria (e.g., popularity, date).')):
    '''Browse available plugins and extensions in the GitHub Marketplace.'''
    typer.echo('Executing browse marketplace command')
    # Examples:
    # gh marketplace browse --category productivity
    # gh marketplace browse --sort popularity

@GitHub Marketplace_app.command(name='install', help='Install a plugin or extensions from the GitHub Marketplace.')
def install(item_name: str = typer.Argument(..., help='Name of the plugin or extensions to install.'), source: str = typer.Option(..., help='Source repository of the plugin or extensions.')):
    '''Install a plugin or extensions from the GitHub Marketplace.'''
    typer.echo('Executing install marketplace command')
    # Examples:
    # gh marketplace install gh-extras/gh-deploy-plugin

app.add_typer(GitHub Marketplace_app, name='GitHub Marketplace')





hygen_app = typer.Typer(help='Integrate Hygen for code scaffolding.')

@hygen_app.command(name='hygen', help='Code scaffolding using Hygen templates.')
def hygen(template: str = typer.Option(..., help='Select template for scaffolding.'), component_name: str = typer.Argument(..., help='Name of the component to scaffold.')):
    '''Code scaffolding using Hygen templates.'''
    typer.echo('Executing hygen command')
    # Examples:
    # gh hygen --template react component Button
    # gh hygen --template node service AuthService

app.add_typer(hygen_app, name='hygen')

assist_app = typer.Typer(help='Integrate LLM-powered code assistants for enhanced code generation and assistance.')

@assist_app.command(name='assist', help='Interact with LLM-powered code assistants.')
def assist(prompt: str = typer.Option(..., help='Provide a prompt for the assistant.'), model: str = typer.Option(None, help='Specify the LLM model to use.'), code_file: str = typer.Argument(None, help='File to apply assistant's suggestions.')):
    '''Interact with LLM-powered code assistants.'''
    typer.echo('Executing assist command')
    # Examples:
    # gh assist --prompt 'Optimize this function' main.py
    # gh assist --prompt 'Generate unit tests' src/utils.py --model gpt-4-turbo

app.add_typer(assist_app, name='assist')




if __name__ == "__main__":
    app()
